requirements.txt
aerosandbox==4.1.1
aix360==0.3.0
Bio==1.5.9
black==23.7.0
boto3==1.28.40
branca==0.6.0
cairosvg==2.7.1
casadi==3.6.3
codecarbon==2.3.1
colorlover==0.3.0
colour==0.1.5
cudf==0.6.1.post1
cufflinks==0.17.3
cuml==0.6.1.post1
cupy==12.2.0
dash==2.13.0
dash_bio==1.0.2
dash_bootstrap_components==1.4.2
dash_canvas==0.1.0
dash_colorscales==0.0.4
dash_core_components==2.0.0
dash_cytoscape==0.3.0
dash_daq==0.5.0
dash_deck==0.0.1
dash_html_components==2.0.0
dash_labs==1.2.0
dash_pivottable==0.0.2
dash_player==1.1.0
dash_slicer==0.3.0
dash_table==5.0.0
dash_vtk==0.0.9
datashader==0.15.2
Flask==1.1.4
flask_caching==2.0.2
floris==3.4.1
folium==0.14.0
gensim==4.3.2
geopandas==0.13.2
grpc==1.0.0
grpcio==1.25.0
holidays==0.31
ipywidgets==8.1.0
itk==5.3.0
joblib==0.14.1
jupyter_dash==0.4.2
Keras==2.3.1
lasio==0.31
lorem==0.1.1
lyft_dataset_sdk==0.0.8
matplotlib==3.1.2
networkx==2.5.1
nibabel==5.1.0
nilearn==0.10.1
nltk==3.8.1
numpy==1.16.5
openai==0.28.0
opencv_python==4.2.0.34
paho_mqtt==1.6.1
pandas==0.25.3
Pillow==10.0.0
plotly==5.16.1
protobuf==4.24.2
PVGeo==2.1.0
pydeck==0.8.0
pydot==1.4.2
pyiso==0.4.0
pymapd==0.26.0
pymongo==3.10.1
pysal==23.7
pyserial==3.5
pytesseract==0.3.10
python_dateutil==2.8.2
pytz==2019.3
pyvista==0.42.0
requests==2.18.4
scikit_learn==1.3.0
scipy==1.3.3
seabreeze==2.4.0
segyio==1.9.11
Shapely==2.0.1
simple_salesforce==1.12.4
six==1.11.0
skimage==0.0
snowflake==0.0.4
spacy==3.6.1
SQLAlchemy==1.3.24
statsmodels==0.14.0
tensorflow==1.14.0
tensorflow_hub==0.14.0
timezonefinderL==4.0.2
timm==0.9.6
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.1
umap==0.1.1
uszipcode==1.0.1
visa==2.21.0
vtk==9.2.6
whitenoise==6.5.0
wordcloud==1.9.2
xarray==2023.8.0
xgboost==1.7.6
